Report strategies created and refactoring done
