+++ model/projectReportType.js
@@ -1,0 +1,9 @@
+const ProjectReportType = {
+    VISUALREPORT : "visualReport",
+    INVASIVEONLY : "invasiveOnly",
+    INVASIVEVISUAL : "invasiveVisual",
+    CONCLUSIVE : "conclusive",
+    SINGLEPROJECT : "singleProject",
+}
+
+module.exports = ProjectReportType;
