+++ service/reportstrategy/reportGenerationStrategy.js
@@ -1,0 +1,7 @@
+class HtmlReportGenerationStrategy{
+    async generateReportHtml(project,sectionImageProperties){
+        throw new Error('generateReportHtml must be implemented');
+    }
+}
+
+module.exports = HtmlReportGenerationStrategy;
